+++ clash-ghc/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog for the [`clash-ghc`](http://hackage.haskell.org/package/clash-ghc) package
 
+## 0.6.6.1
+* Support for `unbound-generics-0.3`
+
 ## 0.6.6 *December 11th 2015*
 * New features:
   * Remove all existing HDL files before generating new ones. This can be disabled by the `-clash-noclean` flag. [#96](https://github.com/clash-lang/clash-compiler/issues/96)